requirements.txt
playwright
pytest
pytest-playwright
python-dotenv
pyyaml
